+++ angadi/angadi/src/main/java/com/angadi/repository/ProductRepository.java
@@ -14,7 +14,7 @@ public interface ProductRepository extends JpaRepository<Product, Integer> {
     public Product findByProductName(String productName);
 
     @Query("SELECT P FROM Product P WHERE P.category = :category AND P.productPrice BETWEEN :min AND :max")
-    public List<Product> findByCategoryAndPriceBetween(@Param("category") String category, @Param("min") Long min, @Param("max") Long max);
+    public List<Product> findByCategoryAndPriceBetween(@Param("category") String category, @Param("min") Integer min, @Param("max") Integer max);
 
     @Query("SELECT P FROM Product P WHERE P.category = :category AND P.productRatings BETWEEN :min AND :max")
     public List<Product> findByCategoryAndRatingBetween(@Param("category") String category, @Param("min") Double min, @Param("max") Double max);
